% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumincglm.R
\name{cumincglm}
\alias{cumincglm}
\title{Generalized linear models for cumulative incidence}
\usage{
cumincglm(formula, time, cause = "1", link = "identity", data, ...)
}
\arguments{
\item{formula}{A formula specifying the model. The left hand side must be a \link[prodlim]{Hist} object. The right hand side is the usual linear combination of covariates.}

\item{time}{Numeric constant specifying the time at which the cumulative incidence or survival probability effect estimates are desired.}

\item{cause}{Character constant specifying the cause indicator of interest.}

\item{link}{Link function for the cumulative incidence regression model.}

\item{data}{Data frame in which all variables of formula can be interpreted.}

\item{...}{Other arguments passed to \link[stats]{glm} such as weights, subset, etc.}
}
\description{
Generalized linear models for cumulative incidence
}
