% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{genset_func}
\alias{genset_func}
\title{Generate a dataset}
\usage{
genset_func(
  n,
  beta1,
  gamma1,
  beta2,
  gamma2,
  beta.cens = c(0, 0, 0, 0),
  cens.rate = 0.2
)
}
\arguments{
\item{n}{Sample size}

\item{beta1}{Coefficients for cause 1}

\item{gamma1}{Shape for cause 1}

\item{beta2}{Coefficients for cause 2}

\item{gamma2}{Shape for cause 2}

\item{beta.cens}{Coefficients for censoring}

\item{cens.rate}{Target censoring rate, used for testing}
}
\value{
A data frame with the simulated data
}
\description{
Simulates a dataset under the competing risks model with two causes
and censoring.
}
