% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{haz11}
\alias{haz11}
\title{Hazard function under the proportional hazards Weibull model}
\usage{
haz11(t, Z, beta, gamma)
}
\arguments{
\item{t}{Time at which to be evaluated (must be >= 0)}

\item{Z}{Covariate matrix}

\item{beta}{Coefficients to be applied to covariate matrix}

\item{gamma}{Shape parameter}
}
\value{
Hazards evaluated at t
}
\description{
Hazard function under the proportional hazards Weibull model
}
