% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{inverse}
\alias{inverse}
\title{Numeric inverse}
\usage{
inverse(fn, min_x, max_x)
}
\arguments{
\item{fn}{Function to invert}

\item{max_x}{Upper limit of the range of fn}

\item{mix_x}{Lower limit of the range of fn}
}
\value{
A function corresponding to the inverse of fn
}
\description{
Numeric inverse
}
