% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infintesimal-jackknife.R
\name{pseudo_rmst}
\alias{pseudo_rmst}
\title{Compute restricted mean survival}
\usage{
pseudo_rmst(fit, tmax)
}
\arguments{
\item{fit}{Survfit}

\item{tmax}{Max time}
}
\description{
Written by Terry Therneau (2019)
}
