% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumincglm.R
\name{pseudo_rmst2}
\alias{pseudo_rmst2}
\title{Compute restricted mean survival}
\usage{
pseudo_rmst2(sfit, jacks, times, tmax, type = "cuminc")
}
\arguments{
\item{tmax}{Max time}

\item{fit}{Survfit}
}
\description{
Written by Terry Therneau (2019)
}
