% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infintesimal-jackknife.R
\name{rmeanglm.infjack}
\alias{rmeanglm.infjack}
\title{Generalized linear models for restricted mean survival}
\usage{
rmeanglm.infjack(formula, time, cause = "1", link = "identity", data, ...)
}
\arguments{
\item{formula}{A formula specifying the model. The left hand side must be a \link[survival]{Surv} object. The right hand side is the usual linear combination of covariates.}

\item{time}{Numeric constant specifying the maximum time at which the mean is restricted.}

\item{cause}{Character constant specifying the cause indicator of interest.}

\item{link}{Link function for the regression model.}

\item{data}{Data frame in which all variables of formula can be interpreted.}

\item{...}{Other arguments passed to \link[stats]{glm} such as weights, subset, etc.}
}
\value{
A pseudoglm object, with its own methods for print, summary, and vcov. It inherits from
glm, so predict and other glm methods are supported.
}
\description{
Using the infintesimal jackknife pseudo observations for the restricted mean
survival, this function then runs a generalized linear model
 The link function can be "identity" for estimating differences,
 "log" for estimating ratios, and any of the other link
functions supported by \link[stats]{quasi}.
}
