% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{sampletimesfunc}
\alias{sampletimesfunc}
\title{Sample event times}
\usage{
sampletimesfunc(mat, b1, g1, b2, g2)
}
\arguments{
\item{mat}{covariate matrix}

\item{b1}{Coefficients for cause 1}

\item{g1}{Shape for cause 1}

\item{b2}{Coefficients for cause 2}

\item{g2}{Shape for cause 2}
}
\value{
A vector of overall survival times
}
\description{
Overall survival times in the competing risks model using the
probability integral transform.
}
