% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate a dataset}
\usage{
simulate_data(n, scenario = "0", beta.cens = c(1, 0, 0, 0))
}
\arguments{
\item{n}{Sample zie}

\item{scenario}{Character identifying the scenario}

\item{beta.cens}{Coefficients for the censoring distribution}
}
\description{
Wrapper for the used scenarios
}
