% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{St}
\alias{St}
\title{Overall survival probability}
\usage{
St(t, Z, beta1, gamma1, beta2, gamma2)
}
\arguments{
\item{t}{Time at which to be evaluated (must be >= 0)}

\item{Z}{Covariate matrix}

\item{beta1}{Coefficients to be applied to covariate matrix for cause 1}

\item{gamma1}{Shape parameter for cause 1}

\item{beta2}{Coefficients to be applied to covariate matrix for cause 2}

\item{gamma2}{Shape parameter for cause 2}
}
\value{
Survival probabilities evaluated at t
}
\description{
Overall survival probability under a competing risks model with two causes
}
