% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{switch_scenario}
\alias{switch_scenario}
\title{Generate survival data for analysis}
\usage{
switch_scenario(scenario, beta.cens = c(1, 0, 0, 0))
}
\arguments{
\item{scenario}{Character constant determining the scenario}

\item{beta.cens}{Coefficient vector for the association with the censoring distribution}
}
\value{
A list with components for each of the coefficient vectors/matrices
}
\description{
Set up the coefficients for generating competing risks data. The coefficients
correspond to the covariate vector with distribution (constant, binomial(.5), N(0,1), N(0,1))
}
