% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate-data.R
\name{true_values}
\alias{true_values}
\title{Get true values of the regression coefficients}
\usage{
true_values(scenario = "0", link = "identity", nchunk = 500, chunks = 100)
}
\arguments{
\item{scenario}{Character string identifying the scenario}

\item{link}{Link function}

\item{nchunk}{Sample size for each chunk}

\item{chunks}{Number of chunks}
}
\value{
A list with the true coefficients for the cumulative incidence and restricted mean
}
\description{
Get true values of the regression coefficients
}
